refactor: 5xx 서버 에러 핸들링 세부화
