+++ src/main/java/gift/service/KakaoAuthService.java
@@ -4,7 +4,7 @@ import gift.config.KakaoProperties;
 import gift.dto.KakaoTokenRequestDto;
 import gift.dto.KakaoTokenResponseDto;
 import gift.exception.KakaoAuthenticationException;
-import gift.exception.KakaoClientError;
+import gift.exception.KakaoApiError;
 import gift.exception.KakaoConnectionException;
 import org.springframework.http.HttpStatusCode;
 import org.springframework.http.MediaType;
@@ -47,12 +47,14 @@ public class KakaoAuthService {
                     .body(body.dtoToFormData())
                     .retrieve()
                     .onStatus(HttpStatusCode::is4xxClientError, (request, res) -> {
-                        KakaoClientError error = KakaoClientError.from(res.getStatusCode());
+                        KakaoApiError error = KakaoApiError.from(res.getStatusCode());
                         String errorMessage = error.getMessage();
                         throw new KakaoAuthenticationException(errorMessage + "응답 코드: " + res.getStatusCode());
                     })
                     .onStatus(HttpStatusCode::is5xxServerError, (request, res) -> {
-                        throw new KakaoConnectionException("카카오 서버에 오류가 발생했습니다. 응답 코드: " + res.getStatusCode());
+                        KakaoApiError error = KakaoApiError.from(res.getStatusCode());
+                        String errorMessage = error.getMessage();
+                        throw new KakaoConnectionException(errorMessage + "응답 코드: " + res.getStatusCode());
                     })
                     .body(KakaoTokenResponseDto.class);
